pyproject.toml
[tool.poetry]
name = "new-project"
version = "0.1.0"
description = ""
authors = ["Viktor <ehl310892@gmail.com>"]
readme = "README.md"
packages = [{include = "new-project"}]

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^2.25.1"
requests = "^2.28.2"
python-dotenv = "^1.0.0"
aiosqlite = "^0.19.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
